# Optimise labels.Compare