@roidelapluie 2020-01-01
can we also microoptimize
if lb := len(b); lb < l {
 l = lb
}
I doubt that it would make a big difference however.


anyway, LGTM.