@brian-brazil 2020-01-01
> I doubt that it would make a big difference however.

It actually makes it slower, which is surprising.